LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/unregister.https-expected.txt


PASS Unregister twice 
FAIL Register then unregister assert_unreached: unexpected rejection: assert_equals: unregistration should resolve with true expected true but got false Reached unreachable code
